Logic complete + fix node movement
